refactor(UploadView): cleanup
