+++ app/packages/core/src/plugins/SchemaIO/components/UploadView/UploadView.tsx
@@ -1,15 +1,12 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React from "react";
 import { Alert, Box, Typography } from "@mui/material";
-import { useFileUpload, createXhrTransport } from "@fiftyone/upload";
 import HeaderView from "../HeaderView";
 import { getComponentProps } from "../../utils";
 import UploadDropZone from "./UploadDropZone";
 import UploadFileList from "./UploadFileList";
 import UploadTotalProgress from "./UploadTotalProgress";
 import UploadActions from "./UploadActions";
-import useFormSync from "./useFormSync";
-
-const transport = createXhrTransport();
+import { useUploadOrchestration } from "./useUploadOrchestration";
 
 export default function UploadView(props) {
   const { onChange, path, schema, data } = props;
@@ -23,23 +20,10 @@ export default function UploadView(props) {
     read_only: readOnly,
   } = view;
 
-  const autoUploadRef = useRef({ destination, endpoint: "/files/upload" });
-  autoUploadRef.current = { destination, endpoint: "/files/upload" };
-
-  const { onFileSuccess, removeFileByPath } = useFormSync({
-    path,
-    data,
-    onChange,
-  });
-
   const {
     files,
     errors,
-    upload,
-    cancel,
     retry,
-    cancelAll,
-    deleteAll,
     dropProps,
     inputProps,
     browse,
@@ -47,56 +31,22 @@ export default function UploadView(props) {
     completedFiles,
     failedFiles,
     isUploading,
-  } = useFileUpload({
-    multiple: true,
+    dropZoneVisible,
+    atFileLimit,
+    handleCancel,
+    handleCancelAll,
+    handleDeleteAll,
+    setAddMoreOpen,
+  } = useUploadOrchestration({
+    path,
+    data,
+    onChange,
+    destination,
     accept,
     maxSize,
+    maxFiles,
     maxConcurrent,
-    transport,
-    onFileSuccess,
   });
-
-  // Trigger uploads manually so we always use the latest destination from
-  // the operator schema, avoiding stale closures from autoUpload memoization.
-  useEffect(() => {
-    if (files.some((f) => f.status === "selected")) {
-      upload(autoUploadRef.current);
-    }
-  }, [files, upload]);
-
-  const [addMoreOpen, setAddMoreOpen] = useState(false);
-  const dropZoneVisible = files.length === 0 || addMoreOpen;
-
-  const fileCountRef = useRef(files.length);
-  useEffect(() => {
-    if (files.length > fileCountRef.current) {
-      setAddMoreOpen(false);
-    }
-    fileCountRef.current = files.length;
-  }, [files.length]);
-
-  const handleCancel = useCallback(
-    async (id: string) => {
-      const file = files.find((f) => f.id === id);
-      if (file?.remotePath) {
-        removeFileByPath(file.remotePath);
-      }
-      await cancel(id);
-    },
-    [files, cancel, removeFileByPath]
-  );
-
-  const handleCancelAll = useCallback(async () => {
-    await cancelAll();
-    onChange(path, []);
-  }, [cancelAll, onChange, path]);
-
-  const handleDeleteAll = useCallback(async () => {
-    await deleteAll();
-    onChange(path, []);
-  }, [deleteAll, onChange, path]);
-
-  const atFileLimit = maxFiles != null && files.length >= maxFiles;
 
   return (
     <Box {...getComponentProps(props, "container")}>
